fixed warnings

+++ src/cruisin/ecs/ecsSystem_physics.cpp
@@ -116,7 +116,7 @@ namespace at3 {
       TrackControls *trackControls;
       state->get_TrackControls(id, &trackControls);
       //Todo: put engine force application *before* the sim update, with wheel update after?
-      for (int i = 0; i < trackControls->wheels.size(); ++i) {
+      for (size_t i = 0; i < trackControls->wheels.size(); ++i) {
         if (trackControls->wheels.at(i).leftOrRight < 0) {
           trackControls->vehicle->applyEngineForce(trackControls->torque.x, trackControls->wheels.at(i).bulletWheelId);
           trackControls->vehicle->setBrake(trackControls->brakes.x, trackControls->wheels.at(i).bulletWheelId);
@@ -291,7 +291,7 @@ namespace at3 {
     physics->rigidBody->setCollisionFlags(physics->rigidBody->getCollisionFlags()
                                           | btCollisionObject::CF_CUSTOM_MATERIAL_CALLBACK
                                           | btCollisionObject::CF_DISABLE_VISUALIZE_OBJECT);
-    physics->rigidBody->setUserIndex((int) -id); // negative numbers for static objects, let's say.
+    physics->rigidBody->setUserIndex(-(int)id); // negative numbers for static objects, let's say. TODO: follow-up
     dynamicsWorld->addRigidBody(physics->rigidBody);
 
     return true;
